Some fixes for vehicle credits for bots. Still not entirely working, bots' charge bar doesn't reset in going warmup->ingame.
git-svn-id: https://localhost/svn/qwta@333 6eeef719-9b92-4d6a-8419-93dac47c4fdd

+++ source/game/botai/BotThreadData.h
@@ -396,7 +396,7 @@ struct clientInfo_t {
 	int						fireSupportChargedUsed;				//mal: set in idPlayer::Think - arty and airstrikes
 	int						deviceChargeUsed;					//mal: set in idPlayer::Think - landmines, etc.
 	int						deployChargeUsed;					//mal: set in idPlayer::Think - deployables.
-	float					vehicleCreditUsed;					
+	int						vehicleCreditUsed;
 
 	int						mountedGPMGEntNum;
